CKRM E16rc1 io controller support.
This version only implements 20 classes of I/O priority. The next release
candidate will support limits, which can then be used as a means to express
proportional sharing between classes (vservers).

The cfq-iosched.c function was hacked by me to work around an issue with
Fedora Core 2 that does not appear in the vanilla 2.6.8.1 kernel.  The
CKRM folks are aware of this and will fix this problem (hopefully in the
next release, too).